for getisstatus,is select someone's orders by userID

+++ src/main/java/com/example/fortry/dto/OrderInfoDTO.java
@@ -2,11 +2,12 @@ package com.example.fortry.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class OrderInfoDTO {
-    private int id;
-    private int userId;
-    private int amount;
+    private Long id;
+    private Long userId;
+    private BigDecimal amount;
 }
