+++ src/main/java/com/example/fortry/entity/User.java
@@ -2,15 +2,16 @@ package com.example.fortry.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 public class User {
-    private int id;
+    private Long id;
     private String username;
     private String password;
     private int age;
-    private long balance;
+    private BigDecimal balance;
     private LocalDateTime createTime =LocalDateTime.now();
 
     public User(String username, String password,int age) {
